requirements.txt
attrs==21.4.0
bcrypt==3.2.2
blinker==1.4
cachelib==0.7.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.1
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.5.1
MarkupSafe==2.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.1
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
